+++ sql/load_sql_context.sql
@@ -210,6 +210,7 @@ select
                             'oid', pc.oid::int,
                             'name', pc.relname::text,
                             'relkind', pc.relkind::text,
+                            'reltype', pc.reltype::int,
                             'schema', schemas_.name,
                             'schema_oid', pc.relnamespace::int,
                             'comment', pg_catalog.obj_description(pc.oid, 'pg_class'),
@@ -234,53 +235,6 @@ select
                                     )
                                 from
                                     directives d
-                            ),
-                            'functions', coalesce(
-                                (
-                                    select
-                                        jsonb_agg(
-                                            jsonb_build_object(
-                                                'oid', pp.oid::int,
-                                                'name', pp.proname::text,
-                                                'type_oid', pp.prorettype::oid::int,
-                                                'type_name', pp.prorettype::regtype::text,
-                                                'schema_oid', pronamespace::int,
-                                                'schema_name', pronamespace::regnamespace::text,
-                                                -- Functions may be defined as "returns sefof <entity> rows 1"
-                                                -- those should return a single record, not a connection
-                                                -- this is important because set returning functions are inlined
-                                                -- and returning a single record isn't.
-                                                'is_set_of', pp.proretset::bool and pp.prorows <> 1,
-                                                'n_rows', pp.prorows::int,
-                                                'comment', pg_catalog.obj_description(pp.oid, 'pg_proc'),
-                                                'directives', (
-                                                    with directives(directive) as (
-                                                        select graphql.comment_directive(pg_catalog.obj_description(pp.oid, 'pg_proc'))
-                                                    )
-                                                    select
-                                                        jsonb_build_object(
-                                                            'name', d.directive ->> 'name',
-                                                            'description', d.directive ->> 'description'
-                                                        )
-                                                    from
-                                                        directives d
-                                                ),
-                                                'permissions', jsonb_build_object(
-                                                    'is_executable', pg_catalog.has_function_privilege(
-                                                        current_user,
-                                                        pp.oid,
-                                                        'EXECUTE'
-                                                    )
-                                                )
-                                            )
-                                        )
-                                    from
-                                        pg_catalog.pg_proc pp
-                                    where
-                                        pp.pronargs = 1 -- one argument
-                                        and pp.proargtypes[0] = pc.reltype -- first argument is table type
-                                ),
-                                jsonb_build_array()
                             ),
                             'indexes', coalesce(
                                 (
@@ -411,6 +365,58 @@ select
                 )
             ),
             jsonb_build_object()
+        ),
+        'functions', coalesce(
+            (
+                select
+                    jsonb_agg(
+                        jsonb_build_object(
+                            'oid', pp.oid::int,
+                            'name', pp.proname::text,
+                            'type_oid', pp.prorettype::oid::int,
+                            'type_name', pp.prorettype::regtype::text,
+                            'schema_oid', pronamespace::int,
+                            'schema_name', pronamespace::regnamespace::text,
+                            'arg_types', proargtypes::int[],
+                            'arg_names', proargnames::text[],
+                            'num_args', pronargs,
+                            'num_default_args', pronargdefaults,
+                            'arg_type_names', pp.proargtypes::regtype[]::text[],
+                            'volatility', pp.provolatile,
+                            -- Functions may be defined as "returns sefof <entity> rows 1"
+                            -- those should return a single record, not a connection
+                            -- this is important because set returning functions are inlined
+                            -- and returning a single record isn't.
+                            'is_set_of', pp.proretset::bool and pp.prorows <> 1,
+                            'n_rows', pp.prorows::int,
+                            'comment', pg_catalog.obj_description(pp.oid, 'pg_proc'),
+                            'directives', (
+                                with directives(directive) as (
+                                    select graphql.comment_directive(pg_catalog.obj_description(pp.oid, 'pg_proc'))
+                                )
+                                select
+                                    jsonb_build_object(
+                                        'name', d.directive ->> 'name',
+                                        'description', d.directive ->> 'description'
+                                    )
+                                from
+                                    directives d
+                            ),
+                            'permissions', jsonb_build_object(
+                                'is_executable', pg_catalog.has_function_privilege(
+                                    current_user,
+                                    pp.oid,
+                                    'EXECUTE'
+                                )
+                            )
+                        )
+                    )
+                from
+                    pg_catalog.pg_proc pp
+                    join search_path_oids spo
+                        on pp.pronamespace = spo.schema_oid
+            ),
+            jsonb_build_array()
         )
 
     )